% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IrishPolDoc}
\alias{IrishPolDoc}
\title{Data to showcase Irish internal party documents}
\format{
A data frame with 145 rows and 4 columns.
\describe{
\item{i}{ID for individual document}
\item{text}{The text of the policy document}
\item{document_name}{The name of the policy document}
\item{party}{The political party to which the document belongs}
}
}
\source{
Created by Jamie McClenaghan by very unscientifically searching Irish political party websites for internal policy documents
}
\usage{
IrishPolDoc
}
\description{
A dataset containing the internal policy documents of Irish political parties
}
\examples{
data(IrishPolDoc)
}
\keyword{datasets}
