% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IrishPolDocData}
\alias{IrishPolDocData}
\title{Data to showcase f**k count}
\format{
A data frame with 145 rows and 4 columns.
\describe{
\item{i}{ID for individual document}
\item{text}{The text of the policy document}
\item{document_name}{The name of the policy document}
\item{party}{The political party to which the document belongs}
}
}
\source{
Created by Jamie McClenaghan by very unscientifically searching Irish political party websites for internal policy documents
}
\usage{
IrishPolDocData
}
\description{
A dataset containing the number of times the word f**k was used in Ted Lasso by Roy Kent.
}
\examples{
data(IrishPolDocData)
}
\keyword{datasets}
